Structure shell and cleanup tailwind usage
