+++ apps/app/src/app/(auth)/handler/[...stack]/HandlerClient.tsx
@@ -1,9 +1,12 @@
 "use client";
 
+import { usePathname } from "next/navigation";
+
 import { useEffect, useState } from "react";
 import { StackHandler } from "@stackframe/react";
+
 import { stackApp, StackProvider } from "~/lib/auth";
-import { usePathname } from "next/navigation";
+
 
 export function HandlerClient() {
   const pathname = usePathname();
