+++ apps/app/src/app/layout.tsx
@@ -1,8 +1,10 @@
 import type { Metadata } from "next";
+
 import { Geist, Geist_Mono } from "next/font/google";
-import "../styles/tailwind.css";
-import "../styles/stack.css";
-import "../styles/globals.css";
+
+import "~/styles/tailwind.css";
+import "~/styles/stack.css";
+import "~/styles/globals.css";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
